[InstallAPI] Break up headers and add common header for TextAPI types
Before it gets too unwieldy, add a common header for all MachO types
that are used across InstallAPI. Also break up the types in
InstallAPI/Frontend. This both avoids circular dependencies and
logically easier to maintain as more functionality gets added.
